implement LocalsRegistry
This is required because Wasm encodes local variables in groups.
This makes it possible to encode u32::MAX local variables in a small and constant amount of binary space.
Therefore we need a way to efficiently cope with this scenario to protect us against interpreter bombs.
This implementation allows to access local variables in this exploited scenario with a worst time complexity of O(log n).
Besides that local variable usages are cached to further minimize exploitation impact.