implements multi-value Wasm proposal
Note that this PR does no real implementation as it turned out that multi-value is already pretty much supported. So the only thing that this PR does is to allow to enable multi-value Wasm proposal support via the wasmi_v1::Config type. Also this PR integrates the Wasm spec testsuite tests for the multi-value Wasm proposal.