implement new LabelRegistry
This new LabelRegistry replaces the old mechanism of creating branching labels, pinning and resolving them. The new system no longer eagerly resolves labels as early as possible but delays resolution to the compilation process where the IR bytecode is transformed into the Exec bytecode. This simplifies a lot of code and avoids having to memorize label users (Vec) during translation.