Resumable Function Calls (#598)
* initial design and commit

This initial commit introduces the initial API surface and design of the resumable calls API for the new `wasmi` engine.

The implementation of the underlying resumable calls has not yet been done but should be fairly straight forward with the existing design.

* fix doc link

* rename ResumableInvocation::call_resumable -> resume

* replace ResumableInvocation::resumable_type with host_func

The host_func API is more powerful, yet decreases the size of the ResumableInvocation type.

* apply rustfmt

* no longer export Stack from engine mod

* initial naive implementation of resumable calls

This naive implementation introduces a lot of duplicated code in the engine that needs to be resolved before we merge this PR.

* add ResumableInvocation::take_stack method

* deduplicate type checking logic

* implement ResumableInvocation::resume partially

* fix doc link

* push frame when host function errors

This is important to allow to resume the execution by simply popping the pushed frame again.

* remove extension trait by Rust API

I forgot this API already existed ... well ...

* remove unnecessary method parameters

EngineResources is a field in EngineExecutor and therefore does no longer need to be passed.

* apply rustfmt

* remove unnecessary generic parameters from engine methods

* add docs to ResumableTrap::host constructor

* experimental: deduplicate engine logic

This experimental commit deduplicates logic between resumable and non-resumable calls in the Engine implementation. This is experimental as this might regress performance of non-resumable calls which would be pretty bad.

* initial implementation of EngineExecutor::resume_func

* fix doc link

* apply rustfmt

* remove TODO for non required docs

* add ValueStack::empty constructor

Using new led to panics at runtime since new constructor checks its inputs which we do not want or need in this particular instance.

* improve Debug impl for ValueStackRef

* fix bug with resuming funcs

The bug was that host function inputs were not properly dropped from the stack before feeding in the arguments that were supposed to replace the missing host results due to host trapping instead. Therefore the stack was getting bigger and bigger with every resumption.

* add initial test for resuming funcs

* remove debug printlns

* remove get_engine parameter

We simply use the provided ctx parameter with which it is also possible to query the engine.

* do not recycle empty dummy stacks

* extend resumable functions test

* add another resumable call test

* rename EngineExecutor::execute_wasm_func_resumable -> execute_wasm_func

* rename EngineExecutor::execute_func_resumable -> execute_func

* update docs of internal Engine methods

* rename ResumableTrap -> TaggedTrap

This name fits better since on the surface it has nothing to do with resumable functions. However, tagging the origin of the trap is important for resumable calls.

* fix doc links

* remove incorrect debug assertion

It is incorrect since host functions are technically able to return Wasm trap codes.

* add note to Func::call_resumable docs

Notifying users that this API is non standard and might not be available in other Wasm engines.

* apply rustfmt