introduce new ConstRef instruction
This new ConstRef instruction replaces the old Const instruction. The advantage of ConstRef is that it is possible to encode it in 3 bytes. This is a trade off since it is probably slower at execution time. To circumvent unnecessary slowdowns in common cases we introduced I32Const24 and I64Const24 instructions earlier.