allocate CompiledFunc prior to translation
Now instead of allocation CompiledFunc to the `wasmi` Engine as the result of validation and translation of a Wasm function, the CompiledFunc is already allocated and initialized instead.
This allows to access the CompiledFunc index for function calls in other translation units which finally provides the possibility to optimize calls to internal functions.
Also this will make it simpler to run function translations concurrently if we ever decide to want this feature.