add function local constant values translation
This way we can represent constant values that cannot be efficiently encoded as registers with negative index. Upon bytecode execution the base register pointer simply refers to the first function local with index 0 where all function local constant values are pre-pended so no special casing needed when accessing those function locals.

If we go down this road we need to adjust wasmi bytecode and translation for a lot of operators and also adjust tons of tests.