Fix copy instruction for many cases (#783)
* add RegisterSpan::iter_u16 method

This is a more efficient variant of RegisterSpan::iter.

* use RegisterSpan::iter_u16 were possible

* minor fix of docs of CopySpan instruction

* add Instruction::CopySpanRev

This commit does not include translation to actually emit CopySpanRev.

* fix bug in encode_local_set

Replaced debug_assert with if-conditional and added explanatory comment.

* disable merging of copies

We disable merging of copies since it is a hard problem to dissect and resolve overlapping copy instructions and in some cases it is even impossible. Having merged copies makes this even harder. We need a new way to encode copies to circumvent this scenario in its entirety.

* assert no overlapping copies after sort

The problem with this debug assert is that the sorting does not guarantee that there are no overlapping copies but just makes it less likely in most common cases. We need a new approach to handle copies to properly fix this case.

* remove unneeded API

* adjust tests for recent changes (no more copy merge)

* fix bug

* remove CopySpanRev

No longer needed.

* fix import warnings

* apply clippy suggestion

* fix cargo doc issue