use Error instead of Trap in Func::call et.al. and host functions
This is a major refactoring that will significantly affect wasmi users unfortunately.
However, there is no better alternative to having a unified Error type when introducing lazy Wasm function compilation during execution.
This requires execution to handle Error which could be a TranslationError due to problems during lazy translation.
This means, Func::call et.al. also need to return Error instead of Trap.
If we want to allow host function calls to call Wasm functions and propagate their result we therefore also need to return the Error type from host functions instead of just a Trap.
This commit handles all of these cases.
For ease of use we introduced Error::as_trap convenience method.
The great thing about this is that a unified Error type is closer to how Wasmtime API looks and feels. So we kinda improved our Wasmtime mirror with this commit.