Improve differential fuzzing (#879)
* collect all exported names in diff fuzz

* refactor and clean up differential fuzzing

* improve panic message formatting

* add fmt and clippy checks to CI for fuzz directory

* use debug formatting for func name

* use early returns

* add regression test case found via fuzzing

* fix new regression test case

The problem was that Wasm instructions such as {i32,i64,f32,f64}.const, local.get and ref.null did not reset the InstrEncoder last instruction thus not preventing invalid instruction fusion of i32.and and i32.eqz.

* run Wasmtime on all fuzzing branches

* shrink test case a bit

* disabled Wasm backtraces for Wasmtime differential backend

* apply rustfmt

* set stack limits for Wasmi (register)

Due to codegen Wasmi (register) might use more stack space for malicious recursive workloads than Wasmi (stack). Wasmtime theoretically suffers from the same problems as Wasmi (register) but can offset a lot of those problems due to its optimizations.

* add comment about custom limits

* diff fuzz: compare global and memory state

This required a major refactoring of the fuzzing code. In all honesty it is a bit ugly but it works. We are going to need another cleanup soon.