Support Wasm custom sections parsing and processing (#1085)
* add Config::ignore_custom_sections

* add custom sections processing to Wasm module parser

* differentiate between CustomSections[Builder]

* add Module::custom_sections iterator getter

* remove unused append method

* optimize memory allocations of custom sections

Now all names and data of all custom sections are stored in a single vector instead of having many small allocations for many custom sections. This is a guard against malicious Wasm inputs with tons of tiny custom sections putting heavy strain on the memory allocator. Also with this implementation Drop can be simpler, only having to deallocate two vectors.

* add test to check if the optimization works

* change docs

* improve documentation

* apply rustfmt