Allow Wasm module instantiation in host functions called from Wasmi's executor (#1116)
* add host call instantiation test

This test currently dead locks as expected.

* add len_params and len_results to HostFuncEntity

* replace uses of HostFuncEntity::ty_dedup

* no longer use FuncTypeRegistry in Wasmi executor

This finally resolves the dead lock when instantiating a Wasm module in a called host function.

* add better HostFuncEntity::new2 constructor

* use new HostFuncEntity::new2 constructor

* remove old HostFuncEntity::new constructor

* rename HostFuncEntity::new2 -> new

* remove commented out line of code

* enforce max params and results per FuncType

New enforced maximum parameter and result types per FuncType is 1000. This fits nicely in a u16 value which we will make use of later in the API.

* return u16 instead of usize in FuncType::len_{params,results}

This also make use of u16 in HostFuncEntity to inline cache len_{params,results} as an space optimization.