Rename `CompiledFunc` to `EngineFunc` (#1124)
rename CompiledFunc to EngineFunc

This avoid confusion with `EngineFunc`s that have been initialized lazily and are not _yet_ compiled. `EngineFunc` is a better name because it simply states that the function is stored inside the Wasmi `Engine` which is always true for them independent of their current compilation state.