re-design branch table instructions
The new encoding scheme requires an additional Instruction::BranchTableCopy and Instruction::BranchTableTarget.
The Instruction::BranchTableCopy handles cases that require values to be copied around or returned upon taking a branch whereas the old Instruction::BranchTable now only handles those cases that do not require to copy or return values.

Encoding also got simpler: both branch table instructions are now followed len_targets times by Instruction::BranchTableTarget and in the case of Instruction::BranchTableCopy it is furthermore followed by any of the available copy instructions.

The follow-copy instruction is not executed directly during execution but just holds information for all the branch table targets and their required copies or returns.

This new encoding has the advantage of being simpler to understand (imo), to be more efficient during execution and also not redundant with respect to value copy and return information.

With the help of the Unalign thin-wrapper we were able to uphold 8 bytes size for Instructions.