improve encoding of wrapping istore instructions
This gets rid of unnecessary allocated function local constants for the immediates of wrapping istore instructions by integrating the wrapping in the translator. Before this commit the translator did not wrap the `value` to store before trying to use it and eventually convert it into a function local constant. For instructions such as i32_store8 (and more) this will never again allocate function local constants now because the wrapped i8 always fits in the `Instruction::I32Store8` instruction's `value` field, etc..