fix querying of memory64 in translator
The incorrect behavior was that memory64 was true for all memories if the flag was enabled in the Config. This is nonsense since memory64 allows memories to be of either 32-bit or 64-bit and does not turn all memories into 64-bit. Thus, now the translator queries the particular memory type which is the correct behavior.