refactor how ptr+offset is calculated
In store_at or load_at instructions and variants we used a ptr value of 0 and turned the 32-bit address into the offset. While this worked it was a bit confusing. So from now on we turn the 32-bit address into the ptr value and use an offset with value 0 instead.

This way we no longer have to convert Address32 to Offset64 which was confusing anyways. Also we need a new Offset64::zero constructor.