remove overzealous check in Memory::new
The removed check asserted that the max size of a memory (if present) did not exceed the system limits. However, this is not necessary to check since Wasm memories are grown lazily and thus creation of these memories is allowed as long as they do not grow beyond system limits.