fix memory.init memory64 implementation
The issue was that Wasmi translation handled all 3 parameters to memory.init as i64 type values just like it has to handle all the other memory instructions. However, memory.init is special in that its len and data_src parameters remain i32 types. This is probably because memory64 allows for 64-bit memories and tables but not for 64-bit data and element segments. As of now I do not have an explanation or rational for this but it is what it is.