Add support for the Wasm `memory64` proposal (#1371)
* add memory64 support to `wasmi::Config`

* add memory64 spec tests

* rename memory64 test cases

* add MemoryTypeInner internal type

This allows to share code and state between MemoryType and MemoryTypeBuilder. The type itself is internal to the memory submodule and not part of the public API of the memory type.

* use new MemoryTypeInner in MemoryType[Builder]

* add some new memory64 specific methods

* apply rustfmt

* refactor MemoryTypeBuilder::validate method

* MemoryType subtyping now checks index type

* remove MemoryEntity::size field and use `u64`

* support memory64 in wasmparser::MemoryType conversion

* adjust memory.grow execution handler

* make table.grow execution handler compile for now

* adjust memory.copy execution handler

* adjust memory.fill execution handlers

* adjust memory.init execution handlers

* adjust Wasmi IR memory instructions

* update memory.init translation tests

* adjust translation tests for memory.copy

* adjust translation tests for memory.fill

* make C-API compile again

* fix bug in MemoryTypeBuilder::validate

* make FuncTranslator memory64 aware

* more explicit feature usage in Wasm spec testsuite

* fix querying of memory64 in translator

The incorrect behavior was that memory64 was true for all memories if the flag was enabled in the Config. This is nonsense since memory64 allows memories to be of either 32-bit or 64-bit and does not turn all memories into 64-bit. Thus, now the translator queries the particular memory type which is the correct behavior.

* no longer explicitly disable memory64 in translation tests

This was only done to fix bugs with the old and incorrect memory64 query behavior.

* improve panic message

* make MemoryType accept proper limits for memory64

* implement new load store instruction encoding

This only implements the new load store instruction encoding for the Wasmi translator. The Wasmi executor and all the translation tests still need to be updated and adjusted.

* update UntypedVal load store handlers

* update Wasmi executor load instruction handlers

* pull fetch_reg_and_imm32 into instrs.rs module

This helps deduplicating the implementation logic between load and store execution handlers.

* fix: use delta parameter

* add support in executor store instruction handlers

* remove old effective_address and rename new one

* use MemoryTypeInner::page_size (dedups code)

* rename Offset64::new to split

* replace Const32<u64> with Address32 in Wasmi IR

This improves type safety of the Wasmi IR and makes it a bit more readable in some places as well as making usage intend clearer.

* remove unused From impl

* refactor how ptr+offset is calculated

In store_at or load_at instructions and variants we used a ptr value of 0 and turned the 32-bit address into the offset. While this worked it was a bit confusing. So from now on we turn the 32-bit address into the ptr value and use an offset with value 0 instead.

This way we no longer have to convert Address32 to Offset64 which was confusing anyways. Also we need a new Offset64::zero constructor.

* remove Offset64::zero constructor

This wasn't really needed and also wansn't a big enough win in readability to hold on to it.

* add APIs to query load/store parameters type safely

This allows us to remove From<u32> for Offset64Hi thus improve type safety for user code.

* add Offset16 type to replace Const16<u64> usage

* add unit test for Offset64::{split,combine} methods

* add more test values to unit test

* fix bug in fetch_optional_memory

This now correctly bumps ip by 1 instead of by delta which caused the bug.

* fix bug in memory.grow

* remove overzealous check in Memory::new

The removed check asserted that the max size of a memory (if present) did not exceed the system limits. However, this is not necessary to check since Wasm memories are grown lazily and thus creation of these memories is allowed as long as they do not grow beyond system limits.

* add explaining comment

* make memory.grow check more generic

* bail out if memory.grow exceeds system limits

* apply clippy suggestions

* fix existing load translation tests

* change import formatting

* move offset16 method to top of file

* add new load reg and reg_mem0 tests for memory64

* clean up new tests and test infra

* extend more load translation tests

* update more load_at translation tests

* update load::at_overflow[_mem0] translation tests

* add new memory64 load translation tests

* deduplicate mem0 load translation test impls

* add docs to the MemIdx type

* improve macro hygiene a bit

* make existing store translation tests compile and work again

* apply clippy suggestion

* fix broken doc links

* improve imports formatting

* move utilities from load to parent submodule

* use &self instead of self in IndexType and MemIdx methods

This also allows us to drop Copy and Clone derives in IndexType.

* update, extend and adjust some store translation tests

Now also tests the following
- non-default memory index
- more offset values
- memory32 and memory64

* add comments describing the different sets of test cases

* update and extend existing store[_offset16] translation tests

* update and extend remaining existing store translation tests

* apply clippy suggestion

* remove unused generics

* add new store_at fallback translation tests

* dedup test code

* remove unused trait bound in test code

* addat_imm_fallback store translation tests

* reformat i32_store translation tests

* reformat some i64_store translation tests

* add some more test values to some f{32,64}_store translation tests

* clean up and modernize f32_store translation tests

* update and modernize f64_store translation tests

* replace iterators with for-loop

* rename some store translation test generators

* deduplicate some load translation tests

We merged _mem0 variants with their parents.

* apply rustfmt

* add store_at_imm_fallback translation tests

* dedup some test code

* fix memory.init memory64 implementation

The issue was that Wasmi translation handled all 3 parameters to memory.init as i64 type values just like it has to handle all the other memory instructions. However, memory.init is special in that its len and data_src parameters remain i32 types. This is probably because memory64 allows for 64-bit memories and tables but not for 64-bit data and element segments. As of now I do not have an explanation or rational for this but it is what it is.

* apply clippy suggestions

* move IndexType into its own submodule

This is so we can easily re-use it for TableType later.

* update and adjust TableType, TableEntity and Table types

* update docs of ModuleHeader::get_type_of_memory

* add MemoryType::index_ty getter

This is going to be useful later.

* implement basic table64 support

* fix bug in wasmparser to Wasmi TableType conversion

* apply rustfmt

* apply clippy suggestions

* apply another clippy suggestion

* add Git submodule for memory64 proposal tests

* update memory64 wast tests

* fix bug in MemoryType::dynamic_ty

* add table64 to spectest module

* update memory64 tests

* remove the faulty memory64 submodule

* add new Git submodule with memory64 tests

* use new Wasmi specific memory64 tests

* fix broken doc link

* update wasmi-tests Git submodule to new URL

* remove faulty Git submodule from earlier