make V128 From and {Read,Write}As impls conditional
This is necessary to make `UntypedVal` 64-bit when `simd` is not enabled. We cannot really have those impls with just 64-bit wide `UntypedVal`s. This seems to be the least disruptive design since `ValType` and `Val` always have the same set of variants and `V128` always exists in `wasmi_core`'s crate root.