Add `Val[Type]::V128` to Wasmi crates (#1418)
* re-design V128 constructors

The V128 type now mirrors wasmtime's V128 type with respect to conversions from and to Rust primitives. Mainly this supports conversions between V128 and Rust's u128 primitive.

* remove From<[u8;16]> impl for V128

As wasmtime's V128 type does not have this constructor, we shouldn't have it either. It is trivially replaceable with From<u128> impl for V128.

* add V128 conversions with UntypedVal and TypedVal

* add V128 variant to wasmi_core::ValType

* adjust wasmi crate for ValType changes

* add simd crate feature to wasmi_cli

* adjust wasmi_cli for ValType changes with simd

* add simd support for wasmi_fuzz and fuzz crates

While simd is supported by Wasmi fuzzing it still is disabled. So no actual Wasm simd fuzzing will take place. This is because the wasmi crate does not yet fully support the Wasm simd proposal.

* adjust Wasmi's C-API crates

* make Val[Type] always include V128 variants

This makes it a lot simpler to maintain dependencies to wasmi_core and also fixes some compilation problem with feature resolution in the workspace. Also this forces fewer `cfg` annotations across dependencies and the codebase as a whole.

* make V128 From and {Read,Write}As impls conditional

This is necessary to make `UntypedVal` 64-bit when `simd` is not enabled. We cannot really have those impls with just 64-bit wide `UntypedVal`s. This seems to be the least disruptive design since `ValType` and `Val` always have the same set of variants and `V128` always exists in `wasmi_core`'s crate root.

* fix imports when building docs for wasmi_ir

* apply rustfmt