Fix some SIMD bugs found by fuzzing (#1429)
* fix panic in i8x16.shuffle translation

This missed a trivial check if the current code path is reachable.

* fix encoding bug in v128.storeN_lane instructions

This only affected instructions with offsets that could not be 8-bit encoded.

* fix bug in v128.shift translation and immediate `rhs`

The error was that in case of an immediate `rhs` shift amount the translation pushed a result and `lhs` back to the stack instead of just `lhs`.