move LocalsRegistry out of the Stack
The remaining parts of the Stack's LocalRegistry is only concerned about the indices of the first local operand occurrences on the Stack and shall be renamed in another commit. This re-design allows to overwrite the types of local operands on the Stack which is beneficial to make operators like `(i64.reinterpret_f64 (local.get 0))` not require a copy instruction.