Add `icu4x` project as user to README (#1593)
* add icu4x user to README

* reduce resolution of icu4x combination mark

* increase resolution to 200pixels width

The icu4x logo with just 120pixels width looked noticibly shabby even in the README, so we increase to 200pixels width.

* make height=64 pixels

This works best for Wasmi's readme as it shrinks logos to height=32. While height=32 source looks very bad, the same image with twice the height (=64) looks very good while being almost the same size as the previous 200pixels width.